Identity API: discard cached tokens within 20 minutes of expiration
Some apps find it inconvenient or impossible to do an async token
fetch at the exact moment that they need a token. For example, if they
are attaching tokens to requests through a webrequest API
handler. These apps need to be able to periodically update a token
that they cache themselves.

This change makes the Identity cache discard tokens that are within 20
minutes of expiration. Apps that need to can poll for tokens on a
relatively long timer (e.g. every 15 minutes), and still be relatively
sure that their tokens will not expire.

BUG=266963
R=fgorski@chromium.org

Review URL: https://codereview.chromium.org/24255010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224728 4ff67af0-8c30-449e-8e8b-ad334ec8d88c