Refactor the setup of Shill*Stubs' default environment.
This change untangles some of the Shill*Stubs' implementation from the setup of a default network environment. In particular the setup function has now access to all Shill*Stubs during this setup.

This also correctly adds stub networks with status 'online' to a profile, so that the respective FavoriteState is created.

BUG=126870
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/24150004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224756 4ff67af0-8c30-449e-8e8b-ad334ec8d88c