[InstantExtended] Investigate crash from check for search results page in TemplateURLRef::ExtractSearchTermsFromURL(..)
I suspect this crash is occuring from an invalid NavigationEntry and that GetURL() gets inlined which is why the crash is occurring much higher up the stack. However, I can't determine a good reason why that would be null. Adding a CHECK to see if we can get more information on why this crash is occurring.

BUG=291348

Review URL: https://chromiumcodereview.appspot.com/24255012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224930 4ff67af0-8c30-449e-8e8b-ad334ec8d88c