Add utm_source tag to webstore URLs loaded from the app launcher.
Clicking on the webstore from the apps grid, or opening the context menu
of an app in the apps grid and clicking "View in Web Store", the web store
should open with "utm_source=chrome-app-launcher" in the URL query parameters.

Clicking on the webstore from the app search results list, or opening the
context menu of an app in the search results and clicking "View in Web Store",
or clicking on a webstore search result, the web store should open with 
"utm_source=chrome-app-launcher-search" in the URL query parameters.

Additionally, installing an app from the app search results should include 
"installsource=applauncher" in the URL query parameters for the extension
download instead of "installsource=inline".

Conflicts:
        chrome/browser/ui/app_list/search/webstore_result.cc

BUG=276653
TEST=Check the URLs under the given conditions. WebstoreInstallerTest unit test
checks that installation URL adds installsource parameter.

Review URL: https://chromiumcodereview.appspot.com/24022002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225127 4ff67af0-8c30-449e-8e8b-ad334ec8d88c