Makes hiding window conditional when escape is pressed in move loop
Hiding makes sense if a new browser window was created for the drag
(as escape=revert, which means we're going to destroy the
window). But if the drag isn't because of a new window (you selected
all the tabs in a window and started dragging), then revert should
move the window backs to its start location and stop.

BUG=297054
TEST=see bug
R=ben@chromium.org

Review URL: https://codereview.chromium.org/24378002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225189 4ff67af0-8c30-449e-8e8b-ad334ec8d88c