Remove screen dimensions from image sent with Feedback. After speaking with the engineers on the feedback backend, I confirmed that the screen dimensions we send are ignored and instead the dimensions are pulled from the image itself. Hence this CL removes the dimensions and replaces them with a placeholder value (so that the protobuf to remain valid - without any dimensions, the server won't accept the report).
This accomplishes two things,
first, the hack we put in place to 'remember' the screen dimensions associated with a feedback report can now be removed.
second, not requiring a browser for dimenions makes it simple to deal with the case where feedback can be invoked without a browser (as referred to in the crbug for this CL).

R=asargent@chromium.org, jamescook@chromium.org, zork@chromium.org
BUG=297334

Review URL: https://codereview.chromium.org/24436002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225275 4ff67af0-8c30-449e-8e8b-ad334ec8d88c