Add Content-Length header to component extension url responses
When investigating the bug referenced below, we found that playing audio via
an <audio> tag from a file inside an extension worked if the extension was
installed normally, but not if it was a component extension. It turns out that
sending a Content-Length header makes this work properly.

BUG=284815

Review URL: https://codereview.chromium.org/23600029

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225335 4ff67af0-8c30-449e-8e8b-ad334ec8d88c