[rac] Use dark grey instead of black color for overlay message
This patch changes the color of the text on the overlay page from
Sk_ColorBLACK (000) to Sk_ColorDKGRAY (444). This change affects all
platforms, because the color is specified in the platform independent
dialog controller.

The patch also reduces the font size by 1 point on the overlay message.

Although the patch attempts to match mocks as close as possible, there're still
slight differences because mocks use a font that's not available in Chrome and
have a lower anti-aliasing value, which makes the font look skinnier than what
is possible in Chrome.

BUG=294485

Review URL: https://codereview.chromium.org/24452002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225426 4ff67af0-8c30-449e-8e8b-ad334ec8d88c