Read PNaCl fingerprint from disk.
This change is redundant given the pending refactoring of the PNaCl
installer, but that change is blocked on resolving the on-demand
feature. In the meantime, we want to read the read the diff update
fingerprint from disk to avoid distributing the wrong delta blobs.

BUG=245318

Review URL: https://codereview.chromium.org/24532003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225490 4ff67af0-8c30-449e-8e8b-ad334ec8d88c