Makes destruction of DesktopNativeWidgetAura saner (2)
Attempt 2 at this. This is same as last patch but I'm disabling unit
test on linux as ownership of X widget is different than HWND when
destroying parent with transients. Will file bug to track it down.

Prior to this patch DesktopNativeWidgetAura destruction keyed off
destruction of the window_ (not the rootwindow). This meant that by
the time we destroyed the Widget the RootWindow was still live. This
meant any children of the rootwindow, such as child widgets, were
alive as well. The general expectation is that by the time a Widget is
destroyed and child Widgets have also been destroyed.

This patch changes destruction so that we key off the
AcceleratedWidget going away. When that happens we destroy the
RootWindow. By doing this we ensure child Widgets have also been
destroyed.

BUG=288953
TEST=covered by test
R=ben@chromium.org
TBR=ben@chromium.org

Review URL: https://codereview.chromium.org/24990003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225814 4ff67af0-8c30-449e-8e8b-ad334ec8d88c