Don't change PrerenderTracker state in WillTransferRequestToNewProcess
Not all of the prerender state was being updated in that function. Instead,
DCHECK that we're not in a prerender. Prerender can't handle cross-process
navigations right now and should abort the prerender before the navigation
occurs. This reverts the PrerenderTracker parts of
https://src.chromium.org/viewvc/chrome?revision=225263&view=revision

Add a browser test which tests that prerenders are aborted correctly. Also
clean up a bit of %-escaping for redirects in the tests.

BUG=298988,106407
TEST=PrerenderBrowserTest.PrerenderCrossProcessServerRedirect

Review URL: https://codereview.chromium.org/24810003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@225987 4ff67af0-8c30-449e-8e8b-ad334ec8d88c