Fixing crash Report - Magic Signature: views::View::ConvertPointToScreen
The problem:
- The BubbleDelegateView tracks the destruction of the widget which owns the anchor view.
- The anchor view gets queried dynamically for its location upon any position change of the anchor widget.
- When a view gets deleted, the widget does not have to be deleted.
Taking that all together and looking at the shelf it is possible to get to the following scenario:

The user puts the shelf into auto hiding mode (so it changes its location). The bubble gets shown. The button - which was used for the anchor - gets deleted, which triggers the destruction of the bubble. The bubble however gets faded out slowly (while the shelf shifts out of view). In this case the location would get queried, but the view is already gone.

BUG=297676
TEST=unittest, but none visually since there isn't a simple reproducible case known

Review URL: https://codereview.chromium.org/24469006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@226158 4ff67af0-8c30-449e-8e8b-ad334ec8d88c