Refactor access to the ONC policy for a network.
Instead of relying on a value stored in Shill (namely, onc_source in UIData) and caching that in NetworkState, directly access the new ONC preference in Chrome to check whether a network is managed or not.
Proxy settings worked already in that way. With this change internet_options_handler uses the same code path.

The only functional change is that a policy indicator will also be visible in the NetworkList of chrome://settings if Ethernet is managed.

BUG=126870
(for API changes; network_state_informer.cc, preferences_browsertest.cc)

R=bartfab@chromium.org, stevenjb@chromium.org
TBR=nkostylev@chromium.org, ygorshenin@chromium.org

Review URL: https://codereview.chromium.org/23609047

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@226237 4ff67af0-8c30-449e-8e8b-ad334ec8d88c