Properly map server prioritites to chrome priorities.
The notification server priorities are different than the chrome priorities.
We should map them so that a server side medium priority is a chrome low
priority, and a server side low priority is a chrome min priority.

The unit tests are also updated for the new priority mappings.

BUG=297255

Review URL: https://codereview.chromium.org/25727004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@226657 4ff67af0-8c30-449e-8e8b-ad334ec8d88c