DevTools: Extract target discovery and manipulation from DevToolsHttpHandlerImpl
Target discovery has moved into DevToolsHttpHandlerDelegate. 

The new DevToolsTarget interface is introduced to encapsulate
access to target details (id, title, url etc) and manipulation 
(activate, close, inspect).

It is now possible to have an implementation of DevToolsTarget
tied to a tab id instead of a RenderViewHost instance (which is the
prerequisite for fixing the referenced bug).

BUG=272174

Review URL: https://codereview.chromium.org/24995003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@226771 4ff67af0-8c30-449e-8e8b-ad334ec8d88c