Parse imageMediaMetadata field from Drive API.
This patch adds parsers for the image media related fields: width, height and rotation of the image. These fields are available in Drive API, but not in WAPI. This is not a big problem, since WAPI is used only for fast fetch. Therefore, these fields may not be available before the full fetch.

TEST=browser_tests
BUG=170309
R=benwells@chromium.org, kinaba@chromium.org, yoshiki@chromium.org

Review URL: https://codereview.chromium.org/25827002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@226973 4ff67af0-8c30-449e-8e8b-ad334ec8d88c