Fix a problem that DRM permission infobar is shown behind the fullscreen surface
When video enters fullscreen, the permission infobar will be shown behind the surface view.
In this case, we should quit the fullscreen video to let user see the infobar.
Also, futher request for protected surface should be ignored.
When user hit the ok button on the permission infobar, we will continue playback the video.

BUG=297639

Review URL: https://codereview.chromium.org/25640007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@227342 4ff67af0-8c30-449e-8e8b-ad334ec8d88c