Store more info about shortcuts, part 1.
This splits the pieces of Shortcut that tracked info about an AutocompleteMatch
into a separate MatchCore struct, which makes it easier to construct Shortcuts
in the various different ways we need, as well as making it simpler to
add/remove future data.

This doesn't actually store any relevant info in the shortcuts database yet, so
there is very little user effect (just during a single run of Chrome).

BUG=304441
TEST=none
R=bartn@chromium.org, brettw@chromium.org

Review URL: https://codereview.chromium.org/26099002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@227576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c