Fix flakyness of MetricsService unit tests.
MetricsService::InitializeMetricsState() has a DCHECK 
that MetricsLog::GetIncrementalUptime() would always 
return 0, which isn't the case for unit tests. (This is 
because MetricsService may be created and destroyed 
multiple times.) 

This CL de-flakes the tests by moving the static 
MetricsLog::GetIncrementalUptime to be a non-static 
MetricsService::GetIncrementalUptime function (with 
the last time stored as an instance variable). 

BUG=290326
TEST=Tests still pass even if you add a sleep() call 
to MetricsServiceTest() constructor.
NOTRY=true

Review URL: https://codereview.chromium.org/25559009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@227767 4ff67af0-8c30-449e-8e8b-ad334ec8d88c