Android: disable 2d canvas AA in a different place.
Instead of disabling 2d canvas antialiasing for Android in
content_startup_flags (which can't easily be overridden by content
embedders), disable it in chrome_startup_flags (which disables it for
Chrome and Chromium TestShell on Android, which is what was desired) and
also in shell_main_delegate (which disables it for ContentShell on
Android for the purposes of consistency in manual/automated testing).
Other content embedders (such as android_webview) are now free to
control this flag themselves.

BUG=213238
R=pfeldman@chromium.org, senorblanco@chromium.org, yfriedman@chromium.org

Review URL: https://codereview.chromium.org/26667005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@227928 4ff67af0-8c30-449e-8e8b-ad334ec8d88c