Makes tabs resize when detaching into a new browser
Prior to this patch when dragging tabs out of a browser window the
tabs in the new browser window were the same size as they were in the
original tabstrip. If the window size changes as part of creating the
new browser (as happens if the window was maximized) this looks
bizarre. New behavior is to resize tabs.

BUG=293698
TEST=see bug
R=ben@chromium.org

Review URL: https://codereview.chromium.org/26574007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@227930 4ff67af0-8c30-449e-8e8b-ad334ec8d88c