[sync] Do not override user-initiated navigations when sync setup is abandoned
When the user abandons an in-progress first-time sync setup by
navigating to another webpage, we automatically sign out the user and
close the existing sync setup overlay. Unfortunately,
OptionsPage.closeOverlay has the side effect of navigating back to
chrome://settings, and this overrides the user-initiated navigation.

This patch adds code to SyncSetupOverlay to check if the sync setup page
is being closed, and if so, to not call OptionsPage.closeOverlay.

BUG=278030
TEST=Abandon first time custom sync setup by navigating away, and make
     sure you aren't redirected to chrome://settings. Also make sure that
     visiting a site like google.com, then going to chrome://settings/syncSetup,
     and then hitting the back button takes you back to google.com.
     to the sync settings page

Review URL: https://codereview.chromium.org/26667003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228011 4ff67af0-8c30-449e-8e8b-ad334ec8d88c