Media Galleries API: Fix settings/content WebUI dialog behavior on add/delete.
Changed the dialog to observe MediaGalleriesPreferences instead of using PrefChangeRegistrar. Previously, the dialog was getting notified of an update before MediaGalleriesPreferences had a chance to update its internal state.

The 'solution' does seem to bloat the SLOC a bit, but word on the street is that this dialog is getting removed soon anyways...

BUG=305079

Review URL: https://codereview.chromium.org/26654005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228025 4ff67af0-8c30-449e-8e8b-ad334ec8d88c