rAc multi-login support
This CL should have no effective change, it's only a change in code to get us closer to supporting multiple users. I've verified that all of this code works with the one change of passing along all the usernames to the account chooser model rather than just the first one.

As there will be some more work required on the view side as well as the server side (probably), this is as good a point as any to stop.

The current problem on the server side is that our passive login flow (sort of) only works for a single user. Multiple users can get passively logged in, but if some do and some do not require passive sign in, we don't handle it well.

BUG=259543

Review URL: https://codereview.chromium.org/25549004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228046 4ff67af0-8c30-449e-8e8b-ad334ec8d88c