Fix update logic of LauncherItem status
There was a problem depending on the navigation timing when two tabs are created and its app is same.
When an inactive tab's loading is finished later than active tab, its LauncherItemStatus is
set to STATUS_RUNNING as a final state by inactive tab's status update.
If the app's item status is active, the item status should not be changed by other inactive tab's
status changing. This problem made two tests(MultipleOwnedTabs andRefocusFilter)flacky.

Also fixed for updating LauncherItemStatus when tab is closing.
When a tab is closed and there is same app's tab is existed, its item should be
updated to running state.

R=skuhne@chromium.org
BUG=303231
TEST=browser_tests

Review URL: https://codereview.chromium.org/25798002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228077 4ff67af0-8c30-449e-8e8b-ad334ec8d88c