Recover version 6 Favicons, fix meta table.
Version 6 is exactly like version 7, except it has an extra trailing
column on favicons which was never used.  The version-7 recovery code
works for it, so long as the meta table is updated correctly.

In writing this change, I realized that the meta table wasn't being
created for recovered databases.  Instead, SQLite was resolving to the
corrupt.meta table.  This works so long as the version doesn't change
(next open will populate meta with kCurrentVersion).  This change
makes the table references absolute, and fixes the meta table setup.

BUG=240396
R=jar@chromium.org, pkotwicz@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/25032004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228217 4ff67af0-8c30-449e-8e8b-ad334ec8d88c