Fix profile-keyed dependencies for custodian profile downloader
The CustodianProfileDownloaderService depends on the
ProfileOAuth2TokenServiceFactory via a ProfileDownloader, so tell the
BrowserContextDependencyManager. Also clean up the ProfileDownloader during the
Shutdown phase rather than letting it sit until the service is destroyed. This
fixes a DCHECK encountered during shutdown when the ProfileDownloader tries to
access a BrowserContext that's been destroyed.

BUG=302436

Review URL: https://codereview.chromium.org/27098004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228660 4ff67af0-8c30-449e-8e8b-ad334ec8d88c