Update defaults for InstantExtended.
Flip the InstantExtended New Tab page to on by default and remove the flags. 
Add a separate flag to control query in the omnibox (which is still off by
default).

Also fixes a ton of tests that needed adjusting for Instant Extended. Mostly
the changes stem from the fact that we're prerendering the Instant Extended
New Tab page which changes assumptions in a bunch of tests.

BUG=297915

Review URL: https://codereview.chromium.org/24733003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228779 4ff67af0-8c30-449e-8e8b-ad334ec8d88c