Password Manager: Make placeholder value the same length as the real password.
In the password manager, the number of dots in each password box should
reflect the actual number of characters in the password. This fixes a
regression introduced by http://crrev.com/226215.

BUG=305139
TEST=Open chrome://settings/passwords, and verify that for any saved password,
the number of dots shown in the password box is the same length as the actual
password.

R=markusheintz@chromium.org

Review URL: https://codereview.chromium.org/26434005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228893 4ff67af0-8c30-449e-8e8b-ad334ec8d88c