[iOS] Hopefully finally fix visibility of Clear Browsing Data button.
Implement an iOS-specific JavaScript solution for hiding the Clear
Browsing Data button when the keyboard is visible. The current media
query-based solution doesn't work in all cases, and it seems that a
pure CSS solution is not possible.

BUG=305740
TEST=Open chrome://history, and check that the Clear Browsing Data
button is never visible when the keyboard is open, for all devices
and orientations.

R=sergiu@chromium.org

Review URL: https://codereview.chromium.org/27528002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@228932 4ff67af0-8c30-449e-8e8b-ad334ec8d88c