This removes the 'keep' option from the download notifications for known malware (dangerous host, URL, or content).
Windows, ChromeOS: This removes the 'keep' option from the dropdown menu that hangs off the notification/warning. Only the 'learn more' option remains in the dropdown menu that hangs off the notification. One exception: mattm is doing a feedback experiment, and if that experiment is turned on there is also a plain 'discard' option in the shelf.

GTK (Mac, Linux): This removes the 'keep' option from the main download notification itself, since there is no dropdown. 

This does *not* change anything in chrome://downloads.

BUG=305387

Review URL: https://codereview.chromium.org/26261010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c