UMA statistics for enabling/disabling synced notification sending services.
For Synced Notifications, we wish to measure how often users turn off
notifications, both in general and for particular sending services.

Due to a limitation in the way that histograms are collected, we an only
collect enum style data.  So, I have chosen to manually represent the first
several sending services.  This obviously doesn't scale as we add more
sending services, so my intent is to add several more, then see if the data
we are getting is still interesting, and perhaps remove the "per-service"
histogram data by the time we have a good set of data.

BUG=280251

Review URL: https://codereview.chromium.org/27267002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229141 4ff67af0-8c30-449e-8e8b-ad334ec8d88c