Identity API: Add chrome.identity.onSignInChanged routing and IDL
This is the first part of adding the chrome.identity.onSignInChanged
event. This change creates the event in dev channel, and an event
router that can dispatch sign-in events to listening extensions.

A new permission, "identity.email" determines whether or not the app
may be given the email address of the account associated with the
user's profile. Apps without the permission still receive events, but
without the email address.

The code to actually generate the events will come in a future CL.

BUG=305830

Review URL: https://codereview.chromium.org/27283002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229345 4ff67af0-8c30-449e-8e8b-ad334ec8d88c