DevTools: Display a warning when trying to open untrusted frontend.
When Chrome is trying to inspect a page on a connected Android device the DevTools frontend is fetched from a URL supplied by the remote browser.
Currently Chrome would crash if this URL is not pointing to a trusted source. For instance, this happens when trying to inspect a page in ContentShell.

This patch removes the CHECKs and displays a text warning instead.

BUG=295979

Review URL: https://codereview.chromium.org/27693002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229371 4ff67af0-8c30-449e-8e8b-ad334ec8d88c