Gets hiding animations to work again
This broke when bounds changes was consolidated into
DesktopNativeWidgetAura::OnRootWindowHostResized. We can't resize the
child layers when this happens as otherwise we're going to force a
paint of the layers and thus the views. But at this point nothing is
valid and we don't want to paint anyway.

Also removing a call to Show() early on as it may trigger animations
that we don't want. A Show() at this time is before the creator has
had a chance to turn them off.

BUG=none
TEST=none
R=ananta@chromium.org, ben@chromium.org, grt@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=229285

Review URL: https://codereview.chromium.org/27043007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229455 4ff67af0-8c30-449e-8e8b-ad334ec8d88c