Find Favicon in priority of icon_type.
This patch provided a way to find the favicon in the order of a list
of icon_types. if the best fit icon is found the rest of icon_types
will not be searched, otherwise, the largest icon of all list
icon_types is returned.

BUG=298446

Review URL: https://codereview.chromium.org/26563004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229760 4ff67af0-8c30-449e-8e8b-ad334ec8d88c