Settings page: Do not expect a reload when navigating between sub-pages
OptionsUI informs the individual handlers whenever the setting page starts
and finishes loading. This was not working correctly when the forward/back
buttons were used to navigate between sub-pages: OptionsUI would inform
the handlers that a load had started but no load is actually needed for
this kind of navigation so that no end of load signal would be generated.

This CL makes OptionsUI wait for the correct trigger so that it signals a
start of load only when a load actually begins.

BUG=301436
TEST=New browser test

Review URL: https://codereview.chromium.org/27042002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229783 4ff67af0-8c30-449e-8e8b-ad334ec8d88c