Use adapter type rather than adapter name in chrome://diagnostics.
Before, chrome://diagnostics assumes that an interface like eth0 will be
an specific type (ie. ethernet). This may not be the case, so we explicitly
check the type of the interface. We use DebugDaemonClient::GetNetworkStatus
instead of DaebugDaemonClient::GetNetworkInterfaces so we can read the type.

BUG=239364
TEST=manual

Review URL: https://codereview.chromium.org/27590003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@229933 4ff67af0-8c30-449e-8e8b-ad334ec8d88c