Add font_scale_factor_quirk pref.
This is an Android-specific boolean pref defaulting to true, which indicates
whether font_scale_factor includes the 1.05-1.3x multiplier that is currently
applied by Clank on startup.

The goal is to move this multiplier out of Clank and into the content layer.
When font_scale_factor_quirk is false, the multiplier is applied in
content::ApplyWebPreferences.

This pref can be removed after Clank code is updated.

BUG=

Review URL: https://codereview.chromium.org/28053002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230042 4ff67af0-8c30-449e-8e8b-ad334ec8d88c