linux_aura: Enable interactive_ui_tests.
Previously, the chrome test runner had a check compiled in only in linux_aura where it detected if the name of the binary was "interactive_ui_tests" and quit immediately.

This removes that check, provides a minimal implementation of ui controls for desktop aura, and enables 56 tests.

BUG=163931

Review URL: https://codereview.chromium.org/26315005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230234 4ff67af0-8c30-449e-8e8b-ad334ec8d88c