Gets hiding animations to work again (3)
This broke when bounds changes was consolidated into
DesktopNativeWidgetAura::OnRootWindowHostResized. We can't resize the
child layers when this happens as otherwise we're going to force a
paint of the layers and thus the views. But at this point nothing is
valid and we don't want to paint anyway.

Also removing a call to Show() early on as it may trigger animations
that we don't want. A Show() at this time is before the creator has
had a chance to turn them off.

Lastly, Widget::ShowInactive() would end up invoking SetFocus() for
DesktopNativeWidgetAura on windows. I'm changing it to not do
that. This broke a chrome_frame test.

BUG=309054
TEST=none
R=ananta@chromium.org, ben@chromium.org, grt@chromium.org
TBR=ananta@chromium.org, grt@chromium.org

Review URL: https://codereview.chromium.org/28923009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230267 4ff67af0-8c30-449e-8e8b-ad334ec8d88c