Files.app: Simplify the parameter type of AppWindowWrapper.
Previously AppWindowWrapper can receive both a parameter object and a function to create the object. But actually, it is enough to specify an object.

Along the way, the CL removes getContentWindows function that is never used now.

This CL simplify the type of parameter and reduces the code.

BUG=310090
TEST=manually

Review URL: https://codereview.chromium.org/35083002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230415 4ff67af0-8c30-449e-8e8b-ad334ec8d88c