Fixed memory leak in icon_manager
When there is an entry in the cache and the |result| is NULL we put NULL in
the cache without deleting current entry. Also I offer to simplify logic
and to dissalow NULLs to reside in the cache. This does not affect any current
code.

BUG=

Review URL: https://codereview.chromium.org/27529002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230614 4ff67af0-8c30-449e-8e8b-ad334ec8d88c