[rAC, OSX] Fix keyView handling.
The rAC dialog used to try to restore whatever the previous keyview was
once the loading shield ended - that triggered inadvertently autopopup
messages.

This change removes the firstResponder hack that did this to hide focus
rings and instead does the right thing - hide the underlying dialog
when the shield is showing.

BUG=310724
R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/39273003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230782 4ff67af0-8c30-449e-8e8b-ad334ec8d88c