Use HighResNow whenever possible on Windows.
This has been enabled on Canary channel for sometime now and seems to be
safe there. We also now have data from stable channel showing there are
no unexpectedly buggy CPUs out there. This should be safe to turn on for
all users with adequate CPUs, which is the majority of them.

BUG=158234

Review URL: https://codereview.chromium.org/41953002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@230968 4ff67af0-8c30-449e-8e8b-ad334ec8d88c