GTTF: Make EmbeddedTestServer always use its own thread for IO
Otherwise we likely end up blocking the browser's IO thread,
especially with big responses that don't fit in the kernel
buffer. This blocks EmbeddedTestServer's IO thread,
and if it's also the browser's IO thread the browser can't
consume the response resulting in a hang.

BUG=310713
R=jam@chromium.org, rch@chromium.org

Review URL: https://codereview.chromium.org/37683004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231075 4ff67af0-8c30-449e-8e8b-ad334ec8d88c