Windows ash own traydelegate
Because the default is not cutting it anymore.

This new delegate is a mix of:
ash\system\tray\default_system_tray_delegate.cc
chrome\browser\chromeos\system\ash_system_tray_delegate.cc

It solves 4 problems in a way similar to chromeos:
1- Upgrade notifications
2- Upgrade nag action --> restart
3- help action --> chrome::ShowHelp()
4- Clock format --> base::GetHourClockType()

For the rest it should be like the default one.

BUG=227247
TEST=none
R=sky@chromium.org, stevenjb@chromium.org

Review URL: https://codereview.chromium.org/30883004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231173 4ff67af0-8c30-449e-8e8b-ad334ec8d88c