First step of new component updater on-demand framework: direct API.
This is the first in a series of CLs to implement the new on-demand
framework for the component update service. I've temporarily
'friended' the PNaCl installer so that I can make those changes in a
separate changelist; once we reach consensus on the triggering
mechanism.

This is a re-upload of issue 27000013.

BUG=307193

Review URL: https://codereview.chromium.org/37943003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231192 4ff67af0-8c30-449e-8e8b-ad334ec8d88c