Fix PrerenderBrowserTestWithExtensions.WebNavigation test.
It got broken after it was disabled. Fix it and bring its disabled configs
in line with other extension browser_test flakes. This is on the theory
that this is the only source of the flake. If the test still fails, it can
be disabled again.

Also remove a chrome.tabs.update call in
PrerenderBrowserTestWithExtensions.TabsApi that wasn't doing anything.

BUG=177163,249179
TEST=PrerenderBrowserTestWithExtensions.*

Review URL: https://codereview.chromium.org/28273007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231451 4ff67af0-8c30-449e-8e8b-ad334ec8d88c