[Net] Make ClientSocketPoolBaseHelper::Group use a timer
Currently it uses a delayed task with a WeakPtrFactory, and checks
the existence of weak pointers to determine whether there is a task
in flight. This is dangerous, since future changes could inadvertently
use the WeakPtrFactory for something unrelated.

Change names to make it clear that a timer is running and
not the backup job itself.

BUG=166689
R=mmenke@chromium.org

Review URL: https://codereview.chromium.org/46053002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231474 4ff67af0-8c30-449e-8e8b-ad334ec8d88c