Force recreating the session service for session restore after shutdown.
Normally, when the session service is shut down, it cannot be recreated
for a profile. This prevents the session service from being recreated to
restore a session and resume session tracking if a browser is re-opened for a
profile without first quitting Chrome.

This change recreates the session service after shutdown if it is to be
used for session restore.

BUG=268224

Review URL: https://codereview.chromium.org/26110003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231528 4ff67af0-8c30-449e-8e8b-ad334ec8d88c