Omnibox: Make BuiltinProvider Play Well with #Hashes
This doesn't appear to me to fix the linked bug (even though Peter
claims it should).  Regardless, it's the right thing to do in any case.

This change also allows BuiltinProvider to respond to omnibox requests
that only want the top match.  (Now that BuiltinProvider can inline,
this is useful.)  It used to respond only to requests that want all
matches.

BUG=280236

Review URL: https://codereview.chromium.org/41583006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c