Files.app: Keep the background page as opened while Files.app's dialog is opening.
Previously Files.app's background page goes to close even if Files.app is
opening as a dialog such as "Open File".  This CL add the flag to background
page and prevent the background page from closing while the dialog is opened.

Along the way, this CL updates the internal process of PathUtil#makeFilesystemUrl
so that it can work even if the background page is gone.

BUG=309970
TEST=manually
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/29683005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231557 4ff67af0-8c30-449e-8e8b-ad334ec8d88c