Convert extensions::LazyBackgroundTaskQueue from Profile to BrowserContext
To do this, introduce the concept of an ExtensionsBrowserClient that allows
the extensions module to ask Profile-like questions of the embedder. The
rest of the change is mostly mechanical.

Also, introduce some conventions around profile() vs. browser_context() and
GetForProfile() vs. GetForBrowserContext() so we can tell which pieces of
the system have been converted.

BUG=309909
TEST=browser_tests LazyBackgroundPageApiTest.* unit_tests LazyBackgroundTaskQueueTest.*

Review URL: https://codereview.chromium.org/46793003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231641 4ff67af0-8c30-449e-8e8b-ad334ec8d88c