Set ChromeViewMsg_SearchBoxSetDisplayInstantResults message param to true only if the underlying page is a results page and the "prefetch_results_srp" flag is enabled via field trials.
Previously, ChromeViewMsg_SearchBoxSetDisplayInstantResults message param was set based on the "prefetch_results_srp" flag value. The underlying page queries this param to enable/disable Instant. Now that we are going to enable Instant behavior outside the search results page, I have to change the condition as follows:

(1) If the underlying page is a results page:
  ChromeViewMsg_SearchBoxSetDisplayInstantResults message param should depend on the "prefetch_results_srp" flag value.
(2) If the underlying page is not a results page :
   ChromeViewMsg_SearchBoxSetDisplayInstantResults message param should depend on the "enable_instant" flag value.

I will submit a follow up CL to address (2).

BUG=269186
TEST=none

Review URL: https://codereview.chromium.org/29253004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231646 4ff67af0-8c30-449e-8e8b-ad334ec8d88c