Mac: delay window raising until space transitions.
When manually raising browser windows in response to a dock activation that
triggered a space transition, isOnActiveSpace is wrong and we focus the wrong
space's windows. Detect this by seeing if the key window is off-screen. If so,
delay the raise until the space change notification.

BUG=309656
TEST=see bug

Review URL: https://codereview.chromium.org/37253004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231647 4ff67af0-8c30-449e-8e8b-ad334ec8d88c