Parse out XSSI guards in Suggestion JSON response
For personalized suggestions in Suggest response, there is a desire for
additional security. In particular, there will be XSSI guards to prevent (well)
XSSI.

We are taking the approach of assuming that XSSI guards will always be there,
and just try to parse it out by looking for the start of valid sugggestion
response. Since the suggestion response is supposed to be a JS array, we look
for "[" in the response. And we repeat the search, in case the XSSI guard
contains "[".

BUG=312458

Review URL: https://codereview.chromium.org/45863006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@231989 4ff67af0-8c30-449e-8e8b-ad334ec8d88c