Adjust the display logic for download destination option for CrOS multi-profile.
To support multi-profiles, the path to the default "Downloads" folder
is going to vary for each profile. We need to generalize the part that
assumed the fixed path "/home/chronos/user/Downloads", and this options
page is one of them.

Some context. The current directory structure looks like:
- /home/chronos/u-xxxxx/Downloads (the primary profile's directory)
- /home/chronos/u-yyyyy/Downloads (the secondary profile's directory)
- ...
- /home/chronos/user (hard link to /home/chronos/u-xxxxx)
and the "user" link points to the primary logged-in profile's directory.
Thus, depending on whether it is loaded as a primary profile or secondary
into a multi-profile session, "/home/chronos/user/Downloads" points to a
different entry. To avoid that, I'm now switching to use the
"/home/chronos/u-xxxxx/Downloads" path always.

BUG=309556

Review URL: https://codereview.chromium.org/50573005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232055 4ff67af0-8c30-449e-8e8b-ad334ec8d88c