Let the progress center remove complete or canceled items immediately.
Previously the progress center remain the complete or canceled item but we
should remove these items to keep the number of items displayed in the progress
center less.

These items actually are not removed from ProgressCenter class until all the
progress items are completed because they are needed to calculate the maximum
progress value. But the view of the items are immediately removed from
ProgressCenterPanel to show a user the less number of items.

BUG=277432
TEST=manually

Review URL: https://codereview.chromium.org/45733011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232090 4ff67af0-8c30-449e-8e8b-ad334ec8d88c