Password Manager should not clean cached passwords on in-page navigation
Some pages (e.g., twitter) do strange main-frame navigation just before submitting a form. We need to prevent this clearing the cached forms. We do this by only clearing the forms if the navigation is not in-page.

This is an alternative solution to https://codereview.chromium.org/40383002/.

BUG=291291, 278130

Review URL: https://codereview.chromium.org/50893002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c