Refactor GuestView to take in extension_id on construction
This is a prerequisite to support relative URL resolution to chrome-extension URLs.

The guest site instance is constructed in BrowserPluginGuestManager::CreateGuest with the appropriate host corresponding to the extension ID of the app/extension hosting the GuestView. This CL gets rid of some code duplication.

BUG=308114
R=creis@chromium.org, jam@chromium.org

Review URL: https://codereview.chromium.org/45313014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232144 4ff67af0-8c30-449e-8e8b-ad334ec8d88c