Keeping the window order between user switches constant
replaced the previous fix for this (activating the windows in the order of the global MRU window list) after a user switch with a show / hide operation in the order as the windows appear in the parent window child list.

The old activation unit test still works and the previously seen layer / windowZ mismatch did not occur anymore.

BUG=311251
TEST=unittest (written previously) & visual control

Review URL: https://codereview.chromium.org/50413014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232265 4ff67af0-8c30-449e-8e8b-ad334ec8d88c