Move Autofill keyboard listening into AutofillPopupControllerImpl
AutofillPopupDelegate is part of the core Autofill code and thus cannot
reference content::KeyboardListener. This CL eliminates keyboard listening
from the AutofillPopupDelegate API and moves the responsibility of deciding
whether AutofillPopupControllerImpl should listen for keyboard input into the
creator of AutofillPopupControllerImpl: the creator should pass
AutofillPopupControllerImpl the WebContents instance in which input should be
listened for, or NULL if it is not desired that the
AutofillPopupControllerImpl listen for keyboard input.

BUG=302499

Review URL: https://codereview.chromium.org/49133003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232448 4ff67af0-8c30-449e-8e8b-ad334ec8d88c