Run a generator to build this resource.
This is a follow-up to https://codereview.chromium.org/23513049/ .

That CL is verifying the presence of expected modules in the browser process, but it relies on checking module name digests that are being read from a resource. That resource was previously manually built, but will now be auto-generated from a list of module names to make it easier to maintain and less error-prone.

BUG=297675
R=jhawkins@chromium.org, robertshield@chromium.org

Review URL: https://codereview.chromium.org/26272002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232459 4ff67af0-8c30-449e-8e8b-ad334ec8d88c