Finish JavaScript side of automated end-to-end tests for extension.
Fix issues that arose (primarily, the security origin needing to be
a valid URL without path, not just a hostname).

Add a message client code can send to determine if the component
extension is available.

Add a way to cancel desktop media chooser in case needed.

Add a way to set upload to default to false, not just to true, in
case needed.

BUG=291271
TBR=nkostylev@chromium.org

Review URL: https://codereview.chromium.org/56033003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232609 4ff67af0-8c30-449e-8e8b-ad334ec8d88c