Disable Background Menu Items for Component Extensions without Options Pages
Component extensions with background that do not have an options page
will cause this menu item to go to the extensions page with an
absent component extension.

Ideally, we would remove this item, but this conflicts with the user
model where this menu shows the extensions with background.

The compromise is to disable the item, avoiding the non-actionable
navigate to the extensions page and preserving the user model.

BUG=311884
TEST=From the Chrome Background Context Menu, "Google Now" should be disabled.
Other extensions (non-component or has options page) will remain enabled.

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=232602

R=kalman@chromium.org, miket@chromium.org, sky@chromium.org, vadimt@chromium.org

Review URL: https://codereview.chromium.org/52553004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232728 4ff67af0-8c30-449e-8e8b-ad334ec8d88c