kiosk: Fix kiosk enable not showing up regression.
When sign in screen shows up for the 2nd time, auto enrollment client already
exists and it is not started to check again. |auto_enrollment_check_done_| is
set to false would stay as false because the done callback is never going
to happen.

BUG=314166
TEST=KioskTest.KioskEnableAfter2ndSigninScreen
R=joaodasilva@chromium.org

Review URL: https://codereview.chromium.org/48993008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c