Record UMA events for actions on the Android NTP.
This records UMA events for actions on the Android NTP (e.g. clicking on
a bookmark), which will be used as part of the time to long dwell
analysis on the NTP. Note: UMA histograms don't contain timestamps, so
we need events instead.

BUG=178932
NOTRY=true
R=mpearson@chromium.org, tedchoc@chromium.org

Review URL: https://codereview.chromium.org/57573002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232753 4ff67af0-8c30-449e-8e8b-ad334ec8d88c