Cleanup: Remove an unnecessary AutocompleteMatch& calculation from OmniboxEditModel::OpenMatch().
I traced the origin of this weird code back to the days before OpenMatch() took
an AutocompleteMatch; back then, we constructed one so we could get at fields
like |fill_into_edit| that weren't provided to OpenMatch().  Now that we pass
this in, I believe the calculated match should always match the passed-in match.

BUG=none
TEST=none
R=sky@chromium.org

Review URL: https://codereview.chromium.org/47353005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232847 4ff67af0-8c30-449e-8e8b-ad334ec8d88c