Add command line flag --disable-webrtc-encryption for dev and canary channels.
It is also be available for "unknown" channel, e.g. unbranded/dev builds. The flag is ignored for beta and stable channels.

The purpose of the flag is for app developer to easier debug RTP media related problems in WebRTC.

https://code.google.com/p/webrtc/issues/detail?id=491

This CL is based on https://codereview.chromium.org/27131004/ by perkj@.

TEST= 1) Start Chrome with the flag --disable-webrtc-encryption. 2) Setup a call using apprtc.appspot.com. 3) Open the JS console and make sure the SDP offer don't contain 'a=crypto:'.
BUG=304024

Review URL: https://codereview.chromium.org/37253002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@232855 4ff67af0-8c30-449e-8e8b-ad334ec8d88c