Deprecate 'singleton' option in chrome.apps.window.create().
This keeps the feature as-is but shows a warning in the developer
console so the developers can fix their code before removing the
feature breaks it.

BUG=313986

TEST=When opening a new window with chrome.apps.window.create() if the
options parameter contain a 'singleton' option set to false and a 'id',
the developer console should show a warning.

Review URL: https://codereview.chromium.org/53973003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233176 4ff67af0-8c30-449e-8e8b-ad334ec8d88c