ActivationTrackerWin: General refactor, renaming and clarification.
I found this class quite difficult to understand, mostly due to bad
method and variable names. Renamed:
  CheckTaskbarOrViewHasFocus -> MaybeDismissAppList,
  RegainNextLostFocus -> ReactivateOnNextFocusLoss,
  regain_next_lost_focus_ -> reactivate_on_next_focus_loss_,
  preserving_focus_for_taskbar_menu_ -> taskbar_has_focus_,
and rewrote many of the comments for clarity.

Factored out most of the logic in MaybeDismissAppList to
ShouldDismissAppList, which returns a bool and has minimal side effects.

This also introduces a minor behavioural fix: If the focused window is
NULL, and reactivate_on_next_focus_loss is true, reactivates instead of
dismissing.

BUG=307837

Review URL: https://codereview.chromium.org/59063009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c