[rAC, OSX] Use bubble windows for error messages.
In the short term, this addresses the issue of messages that are too long
to fit in the dialog - the window will extend outside the dialog. In the
long run, this also prepares for further changes around the error bubble
that do require the BaseBubble machinery.

R=rsesek@chromium.org
BUG=308200

Review URL: https://codereview.chromium.org/49063012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233458 4ff67af0-8c30-449e-8e8b-ad334ec8d88c