Fix some non-aura specific code related to metro startup.
There was a legitimate bug here, but it only occurred on non-aura builds.
On aura builds, the code was still incorrect but by luck exhibited the
correct behavior.

The confusion about this behavior comes from the fact that prior to
r232828, we allowed Desktop and Metro instances of chrome to run 
simultaneously, depending on which shortcut you clicked.  As of r232828, 
all shortcuts lead to the same browser experience, and we store the 
"stickiness" in the registry and the launch experience is based entirely 
off of the value stored in the registry.

As a result, the code here which was incorrect, and which is fixed in this 
patch, only matters for non-Aura.


BUG=314034

Review URL: https://codereview.chromium.org/60823002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233527 4ff67af0-8c30-449e-8e8b-ad334ec8d88c