[rAC, OSX] Add "generated CC" info bubble.
When requestAutocomplete uses Wallet, it generates a card for use for that specific transaction. This code displays an educational bubble, highlighting that it did that and why.

It is implemented as a InfoBubble/BaseBubbleController combo, anchored to a decoration in the location bar. The decoration stays in the bar until the user navigates away, and clicking on it shows/hides the explanation.

This is the cocoa version of it - for the original views code, see http://crbug.com/243574  & https://chromiumcodereview.appspot.com/16467005)
BUG=286527
R=shess@chromium.org, thestig@chromium.org

Review URL: https://codereview.chromium.org/40483003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233719 4ff67af0-8c30-449e-8e8b-ad334ec8d88c