This CL refactors the code for the new password key icon in the omnibox and the password bubble that pops up from the icon. These will replace the password infobar.
The code was previously nested in the code for the content settings but should be changed the functionality needed is too different and it is no content setting in any case.

To refactor this code, some of the code resembles the code used for the zoom icon and bubble.

Only the code for gtk and aura are refactored. The code for cocoa is removed for now and will be added again in another CL.
 
BUG=261628

Review URL: https://codereview.chromium.org/35333002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233726 4ff67af0-8c30-449e-8e8b-ad334ec8d88c