Move AppListModel::Users to AppListViewDelegate
This is some minor re-factoring to simplify AppListModel in preparation
for making it associated with a Profile for syncing. While not strictly
necessary, it is confusing to have a list of "Users" associated with a
model that describes the list of apps for a specific user. Also, the
list of Users is only used by AppListMenu and can be maintained by
AppListViewDelegate.

One side effect of this change is that views will not be notified if the
user list changes and therefore can not invalidate the menu. While this
could be addressed, it shoudln't generally be possible to add a user
while the app list UI is visible, so the added complexity did not seem
worth supporting a theoretical edge case.

BUG=315887
For ash/shell/app_list.cc:
TBR=jamescook@chromium.org

Review URL: https://codereview.chromium.org/63283003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@233996 4ff67af0-8c30-449e-8e8b-ad334ec8d88c