Get rid of user-level styles.
-The Apps codepath for this is just using the wrong thing.
It should be using author-level styles like extensions do.
-The user-stylesheet feature requires the user to put a CSS
stylesheet in the right location in their user-data-dir.
Extensions are a much better way of doing this.

This is in preparation for simplifying the Blink style
resolution code by removing the concept of user styles.

Review URL: https://codereview.chromium.org/64843004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@234007 4ff67af0-8c30-449e-8e8b-ad334ec8d88c